feat(orders): B2B-5157 Rename for file and functions
